Implement path cleaning (#2)
This PR focuses on implementing clean(), a function that takes in a path and simplifies it based on some rules. This is needed as there a could multiple representations of the same path that lead to the same location.

Why not use canonicalize?
Rust has a method called canonicalize which produces the absolute path for any given path pattern. However it raise error when such a path does not exist in the filesystem. However, we still want the user to be able to name new destinations that does not exist in the filesystem.