Ensure order of paths in temporary file
Previously, a HastSet is used to check duplicate paths that user
may input. However, the order of keys in this struct may not be
maintained, resulting in an order which is different from the
order the user inputs in the command prompt.

So instead of using the HashSet in the loop, I decided to clone
the original paths into a vector maintaining the original order.